deleting checkpoints

+++ training/generative_pre_training.py
@@ -72,12 +72,19 @@ def save_checkpoint(model, optimizer, epoch, batch_idx, best_loss, epochs_no_imp
 
     # Delete previous checkpoints
     checkpoints = [f for f in os.listdir(MODEL_STATES_PATH) if f.startswith('checkpoint') and f.endswith('.pth')]
-    if len(checkpoints) > 2:
-        checkpoints.sort(key=lambda x: int(x.split('_')[2].split('epoch')[1]))
-        for old_checkpoint in checkpoints[:-2]: # Keep the last two checkpoints
-            os.remove(MODEL_STATES_PATH + old_checkpoint)
-            if os.path.exists(COLAB_PATH):
-                os.remove(COLAB_PATH + old_checkpoint)
+    if checkpoints:
+        if len(checkpoints) > 2:
+            checkpoints.sort(key=lambda x: int(x.split('_')[2].split('batch')[1].split('.')[0]))
+            for old_checkpoint in checkpoints[:-2]: # Keep the last two checkpoints
+                os.remove(MODEL_STATES_PATH + old_checkpoint)
+             
+    elif os.path.exists(COLAB_PATH):
+        checkpoints = [f for f in os.listdir(COLAB_PATH) if f.startswith('checkpoint') and f.endswith('.pth')]
+        if checkpoints:
+            if len(checkpoints) > 2:
+                checkpoints.sort(key=lambda x: int(x.split('_')[2].split('batch')[1].split('.')[0]))
+                for old_checkpoint in checkpoints[:-2]:
+                    os.remove(COLAB_PATH + old_checkpoint)
 
 # Load checkpoint if exists
 def load_checkpoint():
